#3288 More refactoring
